requirements.txt
Flask==2.2.2
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
google-cloud-bigquery
google-cloud-tasks==2.10.2
